Logic for Admin and SuperAdmin

+++ src/services/Events/DeleteEvents.jsx
@@ -1,13 +1,23 @@
 import axios from "axios";
 import { toast } from "react-toastify";
 
+function getCookie(key) {
+  var b = document.cookie.match("(^|;)\\s*" + key + "\\s*=\\s*([^;]+)");
+  return b ? b.pop() : "";
+}
+
 const DeleteEvents = async (id, reloadReq, setReloadReq, event) => {
   const ServerUrl = process.env.REACT_APP_SERVER_URL;
   try {
-    const res = await axios.delete(`${ServerUrl}/${event}/${id}`);
-    toast.success("deleted successfully");
-    setReloadReq(!reloadReq);
-    return res;
+    const Role = getCookie("CSS_Website_Role");
+    if (Role === "Admin" || (Role === "Executive Head" && event === "abacus") || (Role === "CP Head" && event === "enigma")) {
+      const res = await axios.delete(`${ServerUrl}/${event}/${id}`);
+      toast.success("deleted successfully");
+      setReloadReq(!reloadReq);
+      return res;
+    } else {
+      toast.error("You don't have access to perform this function");
+    }
   } catch (err) {
     toast.error(err.message);
     return err;
